requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==4.1.2
billiard==4.2.0
boto3==1.34.78
botocore==1.34.78
cachetools==4.2.4
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
countryinfo==0.1.2
cron-descriptor==1.4.3
cryptography==42.0.5
CurrencyConverter==0.17.24
daphne==4.1.0
decorator==5.1.1
Deprecated==1.2.14
Django==5.0.6
django-allauth==0.62.1
django-celery-beat==2.6.0
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-filter==24.2
django-js-asset==2.2.0
django-safedelete==1.3.3
django-simple-history==3.5.0
django-storages==1.14.2
django-timezone-field==6.1.0
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
djangorestframework-simplejwt==5.3.1
fabric==3.2.2
facebook-sdk==3.1.0
fake==0.8
Faker==24.4.0
google-api-core==1.22.2
google-api-python-client==2.39.0
google-auth==1.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
iniconfig==2.0.0
invoke==2.2.0
jmespath==1.0.1
kombu==5.3.7
msgpack==1.0.8
oauthlib==3.2.2
packaging==24.0
paramiko==3.4.0
pillow==10.3.0
pluggy==1.4.0
prompt_toolkit==3.0.47
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pytest==8.1.1
pytest-django==4.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytube==15.0.0
pytz==2024.1
redis==5.0.3
requests==2.31.0
rsa==4.9
s3transfer==0.10.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
zope.interface==6.2
